requirements.txt
Flask==3.0.2
line-bot-sdk==3.10.1
google-genai==0.1.0
requests==2.31.0
